feat: apply code suggestions
