--- pallets/unified-accounts/src/impls.rs
@@ -1,175 +1,0 @@
-// This file is part of Astar.
-
-// Copyright (C) 2019-2023 Stake Technologies Pte.Ltd.
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-// Astar is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Astar is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Astar. If not, see <http://www.gnu.org/licenses/>.
-
-use astar_primitives::ethereum_checked::AccountMapping;
-use astar_primitives::evm::EvmAddress;
-use astar_primitives::AccountId;
-use frame_support::traits::OnKilledAccount;
-use frame_support::{pallet_prelude::*, traits::Get};
-use pallet_evm::AddressMapping;
-use precompile_utils::keccak256;
-use sp_core::{Hasher, H160, H256, U256};
-use sp_io::hashing::keccak_256;
-use sp_runtime::traits::{LookupError, StaticLookup, Zero};
-use sp_runtime::MultiAddress;
-use sp_std::marker::PhantomData;
-
-use crate::*;
-
-/// UnifiedAddressMapper implementation
-impl<T: Config> UnifiedAddressMapper<T::AccountId> for Pallet<T> {
-    fn to_account_id(evm_address: &EvmAddress) -> Option<T::AccountId> {
-        NativeToEvm::<T>::get(evm_address)
-    }
-
-    fn to_account_id_or_default(evm_address: &EvmAddress) -> T::AccountId {
-        NativeToEvm::<T>::get(evm_address).unwrap_or_else(|| {
-            // fallback to default account_id
-            T::DefaultAddressMapping::into_account_id(evm_address.clone())
-        })
-    }
-
-    fn to_default_account_id(evm_address: &EvmAddress) -> T::AccountId {
-        T::DefaultAddressMapping::into_account_id(evm_address.clone())
-    }
-
-    fn to_h160(account_id: &T::AccountId) -> Option<EvmAddress> {
-        EvmToNative::<T>::get(account_id)
-    }
-
-    fn to_h160_or_default(account_id: &T::AccountId) -> EvmAddress {
-        EvmToNative::<T>::get(account_id).unwrap_or_else(|| {
-            // fallback to default account_id
-            T::DefaultAccountMapping::into_h160(account_id.clone())
-        })
-    }
-
-    fn to_default_h160(account_id: &T::AccountId) -> EvmAddress {
-        T::DefaultAccountMapping::into_h160(account_id.clone())
-    }
-}
-
-/// AccountMapping wrapper implementation
-impl<T: Config> AccountMapping<T::AccountId> for Pallet<T> {
-    fn into_h160(account: T::AccountId) -> H160 {
-        <Self as UnifiedAddressMapper<T::AccountId>>::to_h160_or_default(&account)
-    }
-}
-
-/// Hashed derive mapping for converting account id to evm address
-pub struct HashedAccountMapping<H>(sp_std::marker::PhantomData<H>);
-impl<H: Hasher<Out = H256>> AccountMapping<AccountId> for HashedAccountMapping<H> {
-    fn into_h160(account: AccountId) -> H160 {
-        let payload = (b"evm:", account);
-        H160::from_slice(&payload.using_encoded(H::hash)[0..20])
-    }
-}
-
-/// AddressMapping wrapper implementation
-impl<T: Config> AddressMapping<T::AccountId> for Pallet<T> {
-    fn into_account_id(evm_address: H160) -> T::AccountId {
-        <Self as UnifiedAddressMapper<T::AccountId>>::to_account_id_or_default(&evm_address)
-    }
-}
-
-/// OnKilledAccout hooks implementation for removing storage mapping
-/// for killed accounts
-pub struct KillAccountMapping<T>(PhantomData<T>);
-impl<T: Config> OnKilledAccount<T::AccountId> for KillAccountMapping<T> {
-    fn on_killed_account(who: &T::AccountId) {
-        // remove mapping created by `claim_account` or `get_or_create_evm_address`
-        if let Some(evm_addr) = EvmToNative::<T>::take(who) {
-            NativeToEvm::<T>::remove(evm_addr);
-            EvmToNative::<T>::remove(who);
-        }
-    }
-}
-
-/// A lookup implementation returning the `AccountId` from `MultiAddress::Address20` (EVM Address).
-impl<T: Config> StaticLookup for Pallet<T> {
-    type Source = MultiAddress<T::AccountId, ()>;
-    type Target = T::AccountId;
-
-    fn lookup(a: Self::Source) -> Result<Self::Target, LookupError> {
-        match a {
-            MultiAddress::Address20(i) => Ok(
-                <Self as UnifiedAddressMapper<T::AccountId>>::to_account_id_or_default(
-                    &EvmAddress::from_slice(&i),
-                ),
-            ),
-            _ => Err(LookupError),
-        }
-    }
-
-    fn unlookup(a: Self::Target) -> Self::Source {
-        MultiAddress::Id(a)
-    }
-}
-
-/// EIP-712 compatible signature scheme for verifying ownership of EVM Address
-/// https://eips.ethereum.org/EIPS/eip-712
-///
-/// Raw Data = Domain Separator + Type Hash + keccak256(AccountId)
-pub struct EIP712Signature<T: Config, ChainId: Get<u64>>(PhantomData<(T, ChainId)>);
-impl<T: Config, ChainId: Get<u64>> SignatureHelper for EIP712Signature<T, ChainId> {
-    type AccountId = T::AccountId;
-    /// evm address type
-    type Address = EvmAddress;
-    /// A signature (a 512-bit value, plus 8 bits for recovery ID).
-    type Signature = [u8; 65];
-
-    fn build_signing_payload(who: &Self::AccountId) -> [u8; 32] {
-        let domain_separator = Self::build_domain_separator();
-        let args_hash = Self::build_args_hash(who);
-
-        let mut payload = b"\x19\x01".to_vec();
-        payload.extend_from_slice(&domain_separator);
-        payload.extend_from_slice(&args_hash);
-        keccak_256(&payload)
-    }
-
-    fn verify_signature(who: &Self::AccountId, sig: &Self::Signature) -> Option<EvmAddress> {
-        let payload_hash = Self::build_signing_payload(who);
-
-        sp_io::crypto::secp256k1_ecdsa_recover(sig, &payload_hash)
-            .map(|pubkey| H160::from(H256::from_slice(&keccak_256(&pubkey))))
-            .ok()
-    }
-}
-
-impl<T: Config, ChainId: Get<u64>> EIP712Signature<T, ChainId> {
-    /// TODO: minor, use hardcoded bytes, configurable via generics
-    fn build_domain_separator() -> [u8; 32] {
-        let mut hash =
-            keccak256!("EIP712Domain(string name,string version,uint256 chainId,bytes32 salt)")
-                .to_vec();
-        hash.extend_from_slice(&keccak256!("Astar EVM Claim")); // name
-        hash.extend_from_slice(&keccak256!("1")); // version
-        hash.extend_from_slice(&(<[u8; 32]>::from(U256::from(ChainId::get())))); // chain id
-        hash.extend_from_slice(
-            frame_system::Pallet::<T>::block_hash(T::BlockNumber::zero()).as_ref(),
-        ); // genesis block hash
-        keccak_256(hash.as_slice())
-    }
-
-    fn build_args_hash(account: &T::AccountId) -> [u8; 32] {
-        let mut args_hash = keccak256!("Claim(bytes substrateAddress)").to_vec();
-        args_hash.extend_from_slice(&keccak_256(&account.encode()));
-        keccak_256(args_hash.as_slice())
-    }
-}
