+++ primitives/src/evm.rs
@@ -57,3 +57,22 @@ where
         Ok(())
     }
 }
+
+/// Mapping between Native and EVM Addresses
+pub trait UnifiedAddressMapper<AccountId> {
+    /// Gets the account id associated with given evm address, if mapped else None.
+    fn to_account_id(evm_address: &EvmAddress) -> Option<AccountId>;
+    /// Gets the account id associated with given evm address.
+    /// If no mapping exists, then return the default evm address.
+    fn to_account_id_or_default(evm_address: &EvmAddress) -> AccountId;
+    /// Gets the default account id which is associated with given evm address.
+    fn to_default_account_id(evm_address: &EvmAddress) -> AccountId;
+
+    /// Gets the evm address associated with given account id, if mapped else None.
+    fn to_h160(account_id: &AccountId) -> Option<EvmAddress>;
+    /// Gets the evm address associated with given account id.
+    /// If no mapping exists, then return the default account id.
+    fn to_h160_or_default(account_id: &AccountId) -> EvmAddress;
+    /// Gets the default evm address which is associated with given account id.
+    fn to_default_h160(account_id: &AccountId) -> EvmAddress;
+}
